requirements.txt
luigi
numpy
pytest
tensorflow-gpu
numpy
SQLAlchemy
